feat(SegmentedSwitch): migrate to Tailwind
